io_uring_sq_ready: always used shared ring state
If we don't use the shared ring state, we can run into issues if a
submit doesn't fully submit. An example of that would be trying to
submit N SQEs, but one of them has an error, so we only partially
submit. But we already flushed the SQ ring to the kernel side of
things, hence subsequent io_uring_sq_ready() checks return 0, even if
we have pending IO to submit in the SQ ring.

Fixes: https://github.com/axboe/liburing/issues/92
Signed-off-by: Jens Axboe <axboe@kernel.dk>