Add io_uring_sqring_wait()
Only applicable when using SQPOLL - allows the caller to wait for space
to free up in the SQ ring, which happens when the kernel side thread has
consumed one or more entries. If the SQ ring is currently non-full, no
action is taken. Note: may return -EINVAL if the kernel doesn't support
this feature.

Signed-off-by: Jens Axboe <axboe@kernel.dk>