test: find free port choice
We have two tests that use a hard wired 9100 as the port, this will
obviously fail if that port is in use.

We really need helpers for this - for now, just find a port by
searching the next 100 ones. If that fails, we skip the test instead of
hard failing it.

Signed-off-by: Jens Axboe <axboe@kernel.dk>