src/queue: never clear data->wait_nr
We're factoring in how many completions we have already, both in
liburing and in the kernel. Hence there's no need to ever clear
it, and in fact it causes issues with early return where we don't
have the required number of CQEs available.

While in there, get rid of the silly 'to_wait' variable.

Signed-off-by: Jens Axboe <axboe@kernel.dk>