test/connect: fix random EACCES for non-root test
We found random failures on `test/connect`, the bug reproduction is very
easy, just run the test in a loop as non-root user and we will see
couple of failures with `EACCES` on bind().

  ammarfaizi2@integral:~/project/now/liburing$ for i in {1..1000}; do test/connect; done;
  bind(): Permission denied
  test_connect(): failed
  bind(): Permission denied
  test_connect(): failed
  bind(): Permission denied
  test_connect(): failed
  bind(): Permission denied
  test_connect(): failed
  bind(): Permission denied
  test_connect(): failed
  bind(): Permission denied
  test_connect(): failed
  bind(): Permission denied
  test_connect(): failed
  ^C
  ammarfaizi2@integral:~/project/now/liburing$

It turned out that we didn't take care of endianess. Fix this by
applying `htons()` after generating port.

We also changed the s_addr assignment with inet_addr("127.0.0.1")
instead of 0x0100007fU to make it cleaner.

Fixes: c8291a94ddf695447f94728a2b819550138e36e8 ("Add a simple connect test")
Reported-by: Louvian Lyndal <louvianlyndal@gmail.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gmail.com>