test: add basic time measurement
Will show the runtime of the test, as well as the previous runtime if
available:

Running test iopoll /dev/nvme0n1p2:			 13 sec [14]
Running test iopoll /dev/nvme1n1p1:			 14 sec [13]

Not perfect yet, and alignment isn't that pretty either. But it's very
useful to see if there are differences between two runs, if you run with
a new kernel that has changes.

We can always pretty it up later.

Signed-off-by: Jens Axboe <axboe@kernel.dk>