src/register: automatically increase RLIMIT_NOFILE, if necessary
A recent change, that also hit stable, means that any file table
sizing will be under RLIMIT_NOFILE as well. We can just adjust this
as needed if someone wants to register a fileset and the count is
larger than RLIMIT_NOFILE rlim_cur. If it's beyond rlim_max, then we
return -EMFILE as the kernel would have done.

Reported-by: Victor Stewart <v@nametag.social>
Signed-off-by: Jens Axboe <axboe@kernel.dk>