src/queue: only enter the kernel if both submit and need_enter are true
We just take the submit count, and then unconditionally just call
sq_ring_needs_enter() to ensure that flags are set appropriately. Only
set need_enter for the submit side if sq_ring_needs_enter() is true,
which may not be the case if SQPOLL is used.

This can avoid a system call if we're not waiting on any CQEs, and
the thread is already running.

Reported-by: Olivier Langlois <olivier@trillion01.com>
Link: https://github.com/axboe/liburing/issues/429#issuecomment-926069364
Signed-off-by: Jens Axboe <axboe@kernel.dk>