nolibc: Don't use `malloc()` and `free()` as the function name
Miyasaki reports that liburing with CONFIG_NOLIBC breaks apps that
use libc. The first spotted issue was a realloc() call that results
in an invalid pointer, and then the program exits with SIGABRT.

The cause is liburing nolibc overrides malloc() and free() from the
libc (especially when we statically link the "liburing.a" to the apps
that use libc).

The malloc() and free() from liburing nolibc use hand-coded Assembly
to invoke mmap() and munmap() syscall directly. That means any
allocation from malloc() is not a valid object with respect to the
libc, and free() will also break the app if we use it to free a
pointer from a libc function (e.g., strdup()).

liburing nolibc should not break any libc app.

This renames the malloc() and free() to __uring_malloc() and
__uring_free(). Also, add new inline functions to wrap the malloc()
and free(), they are uring_malloc() and uring_free(). These wrappers
will call the appropriate functions depending on CONFIG_NOLIBC.

Fixes: c81d0ec560d0493e8cc577a70bbe96bd234cc65f ("Add nolibc build support")
Cc: Tea Inside Mailing List <timl@vger.teainside.org>
Reported-by: Miyasaki Kohaku <kohaku.mski@gmail.com>
Tested-by: Miyasaki Kohaku <kohaku.mski@gmail.com>
Co-authored-by: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gmail.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Link: https://lore.kernel.org/r/20220123074230.3353274-1-ammarfaizi2@gnuweeb.org
Signed-off-by: Jens Axboe <axboe@kernel.dk>