arch/generic: Create arch generic syscall wrappers
This is a preparation for refactoring the syscall wrappers.

This creates a new file src/arch/generic/syscall.h. This file
contains libc syscall wrappers for architectures that don't
have arch specific code. In the next patches, we will include
this file from src/syscall.h.

It aims to reduce the usage of #ifdef/#endif that occurs in
every function in src/syscall.h file. Also, it will make the
arch specific code structure cleaner and easier to manage.

Cc: Nugra <richiisei@gmail.com>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Co-authored-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Link: https://lore.kernel.org/r/20220211155753.143698-2-ammarfaizi2@gnuweeb.org
Signed-off-by: Jens Axboe <axboe@kernel.dk>