configure: Fix clang-12 warning `-Wdeprecated`
```
clang-12: warning: treating 'c' input as 'c++' when in C++ mode, this \
behvior is deprecated [-Wdeprecated]
```

The target compile is C++ but the file extension is ".c". Got the
above warning from clang. Fix it by creating a temporary file with
".cpp" extension and use it.

Introduced by commit e2ed0ec0d74855 ("configure: test for presence of
C++ compiler").

Fixes: e2ed0ec0d74855ca5bce8043d2235beece57aa21 ("configure: test for presence of C++ compiler")
Cc: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>