queue: provide io_uring_get_sqe() symbol again
With a recent change, io_uring_get_sqe() was moved to the header file
to avoid a function call for this hot function. However, this will break
updates of the library, where the symbol is then not provided:

./app: symbol lookup error: ./app: undefined symbol: io_uring_get_sqe, version LIBURING_2.0

Provide one internally so we still export it, but allow newly compiled
applications to grab the one from the header. That should provide the
best of both worlds.

Fixes: c7bf02a38da9 ("Move io_uring_get_sqe() inline in the liburing.h header")
Signed-off-by: Jens Axboe <axboe@kernel.dk>