Revert "make: let src/Makefile set *dir vars properly"
This reverts commit 7551fe972e4b5b811ebbbdf6eef2fbeb2adf3f0a.

"make install DESTDIR=..." specifies a root directory where files are
installed. For example, includedir=/usr/include DESTDIR=/a should
install header files into /a/usr/include.

Commit 7551fe972e4b removed the includedir=, etc arguments on the make
command-line in ./Makefile, leaving only prefix=$(DESTDIR)$(prefix). It
claimed "prefix suffice for setting *dir variables in src/Makefile" but
this is incorrect. "make install DESTDIR=..." now has no effect and
files are not installed with a DESTDIR prefix.

The GNU make manual 9.5 Overriding Variables says:

  all ordinary assignments of the same variable in the makefile are
  ignored; we say they have been overridden by the command line
  argument.

This explains why it was necessary to set includedir=, etc on the make
command-line in ./Makefile. We need to override these variables with
DESTDIR from the command-line so they are not clobbered in src/Makefile
when config-host.mak is included.

Cc: Paymon MARANDI <darwinskernel@gmail.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
Link: https://lore.kernel.org/r/20220414063651.81341-1-stefanha@redhat.com
Signed-off-by: Jens Axboe <axboe@kernel.dk>