configure: fix compile-checks for statx
This is related to the the issue reported here:

https://github.com/axboe/liburing/issues/578

A fix for the reported issue changed ``linux/stat.h`` to ``sys/stat.h``
in ``test/statx.c``, however, it did not change the compile-check in
``configure``.

This change is needed on Alpine Linux / muslc, where 'statx' is detected
by ``configure/compile-check`` but ``sys/stat.h`` does not provide what
is needed by ``test/statx.c``. Resulting in a build-error.

This changes the compiler-check to match the usage in `tests/statx.c`,
thereby fixing the build-error on Alpine Linux / muslc.

Signed-off-by: Simon A. F. Lund <os@safl.dk>