arch/syscall: Use __NR_mmap2 existence for preprocessor condition
Now __NR_mmap2 is only used for i386. But it is also needed for other
archs like ARM and RISCV32. Decide to use it based on the __NR_mmap2
definition as it's not defined on other archs. Currently, this has no
effect because other archs use the generic mmap definition from libc.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Acked-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Link: https://lore.kernel.org/r/20220726111851.3608291-1-alviro.iskandar@gnuweeb.org
Signed-off-by: Jens Axboe <axboe@kernel.dk>